[SofaKernel] Set "dt", "time", "animate" to readOnly in COntext.cpp
I see no meaning full use of changing in the GUI the current "time" for
a node and I'm not even sure could lead to valid simulations.

If one day, someone makes a asyncrhonous/multitime animation loop and use
the GUI then we will reconsider and switch back to non readOnly variables.