[SofaKernel] Remove getOwner, getData, findDataLinkDest, addLink, getName from DDGNode
These function are useless (probably old stuff that was forgotten to remove).
The consequence of keeping them whas a lot of "noise" in the code base and
the object that inherts from DDGNodes have to implement them even if there
is no clear meaning (so a lot returngin nullptr or the empty string).